Moved SongsArrayAdaptor to view package
Seems to me that it make more sense
